% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KidneyCell.R
\name{KidneyCell}
\alias{KidneyCell}
\title{KidneyCell}
\usage{
KidneyCell(
  seurat,
  species = "human",
  method = "Aucell",
  class = 1,
  plot = "heatmap",
  addcelltype = TRUE,
  colors = c("#66C2A5", "#ABDDA4", "#E6F598", "#FFFFBF", "#FEE08B", "#FDAE61", "#D65076")
)
}
\arguments{
\item{seurat}{A Seurat object that has been processed with FindNeighbors and FindClusters.}

\item{species}{Either 'human' or 'mouse', currently only 'human' is supported.}

\item{method}{The scoring method to use, either 'Aucell' or 'Vision', default is 'Aucell'.}

\item{class}{Either 1 or 2, specifying the level of cell type identification. 1 for preliminary, 2 for detailed annotation.}

\item{plot}{The type of plot to visualize the results, either 'heatmap' or 'bubble'.}

\item{addcelltype}{Logical value, whether to add a 'celltype' metadata to clusters, default is TRUE.}

\item{colors}{A vector of colors to use for plotting, default is c("#66C2A5", "#ABDDA4", "#E6F598", "#FFFFBF", "#FEE08B", "#FDAE61", "#D65076").}
}
\value{
A Seurat object with cell type annotations.
}
\description{
This function annotates cell types in a Seurat object based on marker gene scoring, currently supporting human samples only.
}
