% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kidneyDoublet.R
\name{KidneyDoublet}
\alias{KidneyDoublet}
\title{KidneyDoublet}
\usage{
KidneyDoublet(seurat, PC, rate = 8, select = "high")
}
\arguments{
\item{seurat}{A Seurat object to be processed.}

\item{PC}{The number of Principal Components (PCs) to use, which can be obtained from KidneyPC.}

\item{rate}{The rate of doublets per 1000 cells, default is 8.}

\item{select}{Either "high" or "low" to determine the method of doublet removal.}
}
\value{
A filtered Seurat object without doublets.
}
\description{
This function identifies and removes doublets from a Seurat object using the DoubletFinder package.
}
