% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KidneyH5.R
\name{kidneyH5}
\alias{kidneyH5}
\title{KidneyH5}
\usage{
kidneyH5(seurat_obj, output_path)
}
\arguments{
\item{seurat_obj}{A Seurat object to be converted.}

\item{output_path}{Path where the H5AD file will be saved.}
}
\description{
This function converts a Seurat object to an H5AD file format.
}
\details{
This function requires the installation of the reticulate package and the Python anndata library.
It is recommended to run \code{joinlayers} on Seurat v5 objects before using this function.
}
