% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kidneyPC.R
\name{KidneyPC}
\alias{KidneyPC}
\title{KidneyPC}
\usage{
KidneyPC(seurat, reduction = "pca", cum = 90, var = 5)
}
\arguments{
\item{seurat}{A Seurat object that has been normalized, scaled, and processed with PCA or other dimensionality reduction methods.}

\item{reduction}{The dimensionality reduction method to use (e.g., "pca", "harmony"). Must exist in the Seurat object.}

\item{cum}{The cumulative percentage threshold to use for selecting PCs.}

\item{var}{The minimum variation percentage for selecting PCs.}
}
\value{
The number of PCs that meet the given criteria.
}
\description{
This function determines the optimal number of Principal Components (PCs) to use in downstream analysis of a Seurat object.
}
