% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kidneyQC.R
\name{KidneyQC}
\alias{KidneyQC}
\title{kidneyQC}
\usage{
KidneyQC(
  seurat,
  species = "human",
  MT = 5,
  RP = 5,
  HB = 1,
  Feature_high = 5000,
  Feature_low = 200,
  Plot = TRUE,
  color = c("#FF6F61", "#6B5B95", "#88B04B", "#98FB98", "#92A8D1", "#955251", "#B565A7",
    "#009B77", "#F7CAC9", "#D65076", "#45B8AC", "#EFC050", "#5B5EA6", "#9B2335",
    "#DFCFBE")
)
}
\arguments{
\item{seurat}{A Seurat object to be processed.}

\item{species}{Either 'human' or 'mouse', specifying which species the data comes from.}

\item{MT}{The mitochondrial gene filter threshold, default is 5.}

\item{RP}{The ribosomal gene filter threshold, default is 5.}

\item{HB}{The hemoglobin gene filter threshold, default is 1.}

\item{Feature_high}{The upper threshold for RNA feature count, default is 5000.}

\item{Feature_low}{The lower threshold for RNA feature count, default is 200.}

\item{Plot}{Logical value, whether to plot QC metrics (default TRUE).}

\item{color}{A vector of colors to be used in the plot, effective if Plot = TRUE. Default color set is provided.}
}
\value{
A filtered Seurat object after quality control.
}
\description{
This function performs quality control on a Seurat object by calculating the percentages of mitochondrial, ribosomal, and hemoglobin genes and optionally plotting the QC metrics.
}
