% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSeurat.R
\name{scSeurat}
\alias{scSeurat}
\title{scSeurat}
\usage{
scSeurat(h5ad_path)
}
\arguments{
\item{h5ad_path}{A string specifying the file path to the H5AD file.}
}
\value{
A Seurat object with the following components:
\item{counts}{A sparse matrix containing raw or normalized counts.}
\item{meta.data}{A data frame containing cell metadata.}
\item{dimensional reductions}{Dimensionality reduction embeddings (e.g., PCA, UMAP, t-SNE).}
}
\description{
This function converts an AnnData object stored in an H5AD file into a Seurat object.
It supports loading raw counts (if available) or normalized counts and includes metadata
and dimensionality reduction embeddings.
}
\details{
The function first attempts to load raw counts from \code{adata$raw$X}. If raw counts are not available,
it falls back to normalized counts from \code{adata$X}. The function also extracts cell metadata from \code{adata$obs},
gene names from \code{adata$var} (or \code{adata$raw$var} if raw counts are used), and dimensional reduction embeddings from \code{adata$obsm}.

The gene and cell names are added to the counts matrix to ensure compatibility with Seurat.
Dimensional reduction embeddings are added to the Seurat object under their respective names.
}
